Expressions Age
I commented out everything from the begging down to expression of age
because I am going to continue
to use this js file for these lessons so that I may look back at them
later. Also learned how to subtract a year if I did not have a birthday
yet.